Refactoring the function as the previous solution was removing and recreating the dest directory instead of just cleaning it up.
The issue with the previous approach (at least on Linux) is that if you have a terminal open inside the specified directory, and you recreate the directory the terminal process will still try and use the old deleted directory regardless if the new directory matches the same path. As a result, every time directory gets recreated we need to perform change dir operation. This solution eliminates that need because the root dir is never deleted, just it's contents.