Excluding node_modules from being injected in the generated bundle by marking them as external, and doing so by reading "dependencies" section from package.json. Good thing about this is that there is no longer need to deal with better-sqlite3 node bindings. As all dependencies are external. So in the final build we would just copy the whole node_modules folder together with dist when making a distribution artifact.
In addition:

Moved base options from phunt-rollup-config to phunt-cli config as the merging of object properties was shallow so not all keys got integrated from both base file and the main. As I did not want to waste any more time on this decided that for now it is good enough to just move the settings.

Removed rollup terser plugin as I don't see currently the need to minify the generated bundle and it added addional overhead when debugging.

Refs #6